Flatten directory structure
#### Problem

The solana-frozen-abi-macro crate is nested, but it's the only nested
crate in the whole repo.

Also, every crate has a clean mapping between its directory name and
crate name, just requiring a prefix of `solana-`, except for `macro`.

#### Summary of changes

Un-nest `frozen-abi/macro` to `frozen-abi-macro` and rename `macro`
directory to `sdk-macro`.

This will make the crate patch script very easy to write and maintain,
which you will see in the next PR