scripts: Add patching script for Agave devs
#### Problem

Now that the solana-sdk is split out of the Agave repo, it will be
annoying for Agave developers to make simultaneous changes to the SDK
and Agave.

#### Summary of changes

Add a script for patching the solana-sdk in Agave, along with a CI check
to make sure that the SDK doesn't break Agave.