workspace: Relax intra-workspace dependencies
#### Problem

All of the crates within solana-sdk pin their dependency on
intra-workspace crates, ie `solana-program` uses version `=2.2.0` of
`solana-pubkey`. This is a holdover from Agave, where it makes sense to
pin everything to be sure that you're getting exactly what you expect.

Also, since Agave crates do not always adhere to semver, pinning the
local dependencies ensures that the build always works as expected.

Since crates are versioned independently within the solana-sdk repo,
however, the pinning requires us to publish all downstream crates
anytime one of them is updated, to be sure that a build is possible.
This is onerous and unnecessary, since solana-sdk will adhere to semver
going forward.

#### Summary of changes

Relax all of the dependencies to remove the `=`.

Note that we will need to bump all crates to v2.2.1 and re-publish them
so that downstream users can pick up this change, before we can
adequately start publishing crates individually.