workspace: Bump all crates for publish
#### Problem

As mentioned at #27, the sdk crates were pinning their versions
internally, which made them more difficult to use outside of the repo.
Now that the version has been relaxed, we need to publish these crates,
but they're almost all still on v2.2.0.

#### Summary of changes

Normally crate publishing will be done one-by-one through the GitHub
Actions publish workflow, but with ~100 crates, we need to go faster.

Bump all of the v2.2.0 crates to v2.2.1, bump
address-lookup-table-interface to v2.2.2, and bump loader-v3-interface
to v3.0.0 since it contains a breaking change.

Once this lands, we can publish all of the crates in the repo.