program-error: Deprecate PrintProgramError
#### Problem

As part of the breaking change to unite pinocchio's program error with
the sdk's program error in #12, we are introducing a new `ToStr` trait
that doesn't rely on any particular logging.

To go with that, we want to deprecate `PrintProgramError`.

#### Summary of changes

Before landing the breaking change, let's deprecate `PrintProgramError`
properly and add `ToStr` as an alternative for downstream users.

Once this lands, we'll do the patch release with the deprecation, then
we can finally land #12.