secp256k1: Remove libsecp256k1 from public interface
#### Problem

The libsecp256k1 crate used by the secp256k1-program is unmaintained,
but its types appear in the public API of the crate, so we can't upgrade
or change the library without breaking users.

#### Summary of changes

Abstract away the usage of libsecp256k1 by introducing new generic
functions that simply take in byte arrays.