secp256r1: Deprecate openssl types in public API
#### Problem

This is #114, but for secp256r1 instead, removing external types from
the public interface.

#### Summary of changes

This one was a bit more complicated because it uses openssl types, which
aren't as straightforward as simple bytes, so it uses bytes encoded in
DER as the interface. This seemed like the most standard format from
simple web searching, but I'm certainly no expert in this area.

There seemed to also be a lot of vecs in the new instruction
implementation, which shouldn't be necessary, so I changed those to
arrays.