account-info: Rename `rent_epoch` to `unused`
#### Problem

The concept of rent-epoch doesn't exist anymore in the runtime, but the
account-info type for on-chain programs requires some 8-byte field where
the rent-epoch is currently.

#### Summary of changes

Rename the field, skip it in the program entrypoint.

Also, move the layout test into solana-account-info.

BREAKING: `solana_account_info::Account` trait removes `Epoch`