rent-collector: Deprecate everything except the struct
#### Problem

The rent collector is still being used in Agave to hold onto the `Rent`
for the cluster, but is otherwise unused.

#### Summary of changes

Deprecate all functions and constants that should be redefined
elsewhere.

I decided to not deprecate the whole thing since Agave still needs the
struct, but full deprecation is also an option.

This will be backported to maintenance/v2.x. If we decide on full
deprecation, then we can also remove the crate entirely.