hash!: Unify all hash types, introduce "offchain" feature
#### Problem

There are many repeated `Hash` types in all of the different hasher
crates, which brings in unnecessary dependencies and copied code.

#### Summary of changes

Unify all hashes under `solana_hash::Hash`, by re-exporting that type
from keccak and blake3.

While doing this work, I also cleaned up the dependencies, which were
unnecessarily bringing in the crate for the offchain implementation.

When integrating the SDK with Pinocchio, we'll have many situations in
which Pinocchio doesn't want to have an offchain fallback implementation
available, but that fallback is still used by other downstream users.

To satisfy both uses, this PR also introduces a general `"offchain"` crate
feature.  For example, `Pubkey::find_program_address` could use the
offchain feature for deriving the address without using the syscall,
which is useful for most offchain applications.

I could also split off the offchain feature in a different PR if that's
clearer.