wasm-js: Remove system program dependency
#### Problem

While working on the refactor of all wasm-js logic into sdk-wasm-js, I
noticed that no matter what I did, `SystemInstruction` still made it
into the final build. That made it impossible to build the package with
breaking changes.

#### Summary of changes

Remove the usage of `SystemInstruction` in the tests. This required
exposing `Instruction` and `AccountMeta` through wasm-bindgen to be able
to create instructions.

The API doesn't need to be perfect, just usable, so I went with a
builder pattern for adding `AccountMeta`s to an `Instruction`, which
hugely simplifies the logic. Otherwise we'd need to perform conversions
between JS arrays and `Vec<AccountMeta>`, which I was too lazy to do
properly because wasm-js crate is more of a proof-of-concept.

While doing that, I moved the wasm-js deps to the component crates that
were required for the tests.