message: Remove dependency on solana-system-interface
#### Problem

While debugging changes to the wasm-js build, I noticed that
`solana-transaction` and `solana-message` were pulling in
`solana-system-interface` in a lot of situations, even though there's a
lot of code in place to make sure that doesn't happen.

#### Summary of changes

Totally remove the dependency on solana-system-interface by inlining
some nonce-related code in solana-message.