program-error!: Upgrade to v2.2.2 crates (#99)
#### Problem

The repo is still on v2.2.1 sdk crates, but a few versions have been
updated since then. Most importantly, `DecodeError` and
`PrintProgramError` were deprecated in favor of `ToStr`, and
`spl-program-error` implements the old traits.

#### Summary of changes

Update to v2.2.2 crates, and update spl-program-error to stop
implementing `DecodeError` and `PrintProgramError` in favor of `ToStr`.