sdk!: Remove `program` feature
#### Problem

The `program` feature has existed since v1.3 so that people can use
solana-sdk in on-chain programs.

`cargo-build-sbf` even had a special case where it would enable the
`program` feature when building solana-sdk, but it was removed in
https://github.com/anza-xyz/agave/pull/5870.

This means that our current build-sbf script fails when using v2.3 of
the tools.

#### Summary of changes

Remove the `program` feature from the sdk, and fix the build-sbf script
to exclude the sdk, and then try to build without default features.

Rather than adding more excludes, I also changes the build-sbf script to
use `cargo hack`, which is a bit cleaner.