sha256-hasher: Deprecate `extend_and_hash`
#### Problem

As noted in https://github.com/anza-xyz/agave/pull/7028, the
`extend_and_hash` function is unnecessary and creates another
allocation, so we inlined it in Agave.

#### Summary of changes

Deprecate it formally in the crate. This PR will be backported to v2.x
and published, so we can remove the function in a follow-up PR.