deps: Use k256 instead of libsecp256k1
#### Problem

As noted in https://github.com/anza-xyz/agave/pull/4744 and #241,
there are a few issues with libsecp256k1, mainly that it's unmaintained
and has old dependencies, which causes some builds to break.

#### Summary of changes

Port everything over to `k256` instead. It has a lot of similar
interfaces, but I wasn't sure about certain functions, such as
`to_encoded_point(false)`, and whether we're using the message hash
correctly.

To bring more confidence, I expanded one of the doctests to actually run
the code.