msg: Use dependency without default features
#### Problem

The solana-msg crate is brought in with default features, which makes it
impossible to make downstream users of the crate no-std.

#### Summary of changes

Add `default-features = false` for solana-msg at the workspace-level,
and add the std feature to solana-program, which is the only place that
requires it.

As part of this, I noticed that solana-program-entrypoint re-exports
msg, but the re-export is never used, so I removed it.