system-interface: Make it no-std compatible
#### Problem

We're about to publish a breaking change to system-interface, so why not
have a little fun and make it no-std beforehand!

#### Summary of changes

Add alloc feature for vec / string and std feature for frozen-abi.
There's not a ton you can do with the crate currently since the
instructions require bincode / alloc, but we should be able to trim this
more in the future with pinocchio-compatible types.

While testing this, I noticed that the frozen-abi tests didn't work for
no-std crates, which will likely become a problem in the future, so I
fixed the macro to properly emit eprintln rather than through the hacky
way it's done now.