instruction: Forward compatibility for BorshIoError serde
#### Problem

As noted in #404, InstructionError in v2 can't deserialize the new
BorshIoError variant in v3 because we removed the string component.

#### Summary of changes

After trying many different things, this seems like the least bad
option, and actually recommended by serde's maintainer for this kind of
situation.

Agave should not be updated to use this, but all clients can pull in the
new crate for their v2 builds to work with v3 RPC nodes.