instruction: Handle string in all deserializers
#### Problem

The fallback logic to optionally deserialize a string with
`BorshIoError` only works with JSON, and doesn't work with binary
formats like bincode.

#### Summary of changes

I tried so many different things, including a custom visitor,
introducing an intermediate type, deserializing just the bytes first,
and a few others not worth mentioning.

Nothing worked in the end, so I've inlined the macro expansion for
`Deserialize`, and added custom code for handling the `BorshIoError`
variant, where we fallback to an empty string if we fail to deserialize
the string.

Since this code is only in the v2 branch, I think we can live with the
ugliness, but I'm certainly open to any other suggestions.