deps: Bump to rand v0.9 where possible
#### Problem

The repo is still on rand v0.8 in a lot of places, but that version
isn't very compatible with rust 2024 due to its usage of `gen` as a
function name, which is becoming a keyword.

#### Summary of changes

Follow https://rust-random.github.io/book/update-0.9.html to perform the
upgrade to rand v0.9 where possible.

Most crypto libraries are still on rand v0.8 / rand_core v0.6, and the
blstrs and ecdsa implementations of `random` aren't as simple as just
taking 32 or 64 random bytes, so I left those back.